Updated File Structure
