+++ include/GraphMatrix.h
@@ -8,36 +8,9 @@
 #include <stack>
 #include <algorithm>
 #include <set>
-
+#include "MatrixRep.h"
 namespace Appledore
 {
-    // Tag structures
-    struct DirectedG
-    {
-    };
-    struct UndirectedG
-    {
-    };
-    struct UnweightedG
-    {
-    };
-    class GraphVertex
-    {
-    public:
-        size_t id;
-
-        static size_t nextId;
-
-        GraphVertex()
-        {
-            id = nextId++;
-        }
-        bool operator<(const GraphVertex &other) const
-        {
-            return id < other.id;
-        }
-    };
-    size_t Appledore::GraphVertex::nextId = 1;
     template <typename EdgeType>
     struct EdgeInfo
     {
@@ -49,7 +22,7 @@ namespace Appledore
 
     // GraphMatrix class template
     template <typename VertexType, typename EdgeType, typename Direction>
-    class GraphMatrix
+    class GraphMatrix : Appledore::MatrixRepresentation
     {
     public:
         GraphMatrix()
